+++ server/package.json
@@ -22,5 +22,9 @@
     "nodemon": "^2.0.14",
     "sequelize": "^6.9.0",
     "sqlite3": "^5.0.2"
+  },
+  "devDependencies": {
+    "jest": "^27.3.1",
+    "supertest": "^6.1.6"
   }
 }
